Fix Rubocop config validation
As part of our CI, we run `validate_config.rb` which loads the
Rubocop config with `RuboCop::ConfigLoader.load_file`. It is
designed to output the error and exit with a non-zero code, so CI
fails, if there are any errors.

At the moment, it does exit with a non-zero code, but it doesn't
successfully output the message, so you can't see what actually
went wrong.

This fixes the code so we successfully output the messages sent to
stdout and stderr, so you can see what went wrong.